Task 3. To be is to be the value of a variable

+++ 0x10-variadic_functions/3-print_all.c
@@ -33,7 +33,7 @@ void print_all(const char * const format, ...)
 				printf("%d", va_arg(ap, int));
 				break;
 			case 'f':
-				printf("%f", va_arg(ap, float));
+				printf("%f", va_arg(ap, double));
 				break;
 			case 's':
 				stmp = va_arg(ap, char *);
